Fixed abort issue with upload

+++ test/e2e/specs/recordedit/data-independent/add/00-recordedit.add.spec.js
@@ -20,6 +20,20 @@ describe('Record Add', function() {
 			    });
 
 				describe("Presentation and validation,", function() {
+
+
+                    beforeAll(function() {
+
+                        var files = tableParams.files;
+                        if (process.env.TRAVIS)   files = tableParams.files.filter(function(f) { if (!f.doNotRunInTravis) return f; });
+                            
+                        files.forEach(function(f) {
+                            var path = require('path').join(__dirname , "/../../" + f.path);
+                            exec("perl -e 'print \"\1\" x " + f.size + "' > " + path);
+                            console.log(path + " created");
+                        });
+                    });
+
                     var params = recordEditHelpers.testPresentationAndBasicValidation(tableParams);
 				});
 
@@ -92,7 +106,7 @@ describe('Record Add', function() {
 
                                 // if there is a file upload
                                 if (tableParams.files.length) {
-                                   browser.wait(EC.invisibilityOf($('.upload-table')), tableParams.files.length ? (tableParams.records * tableParams.files.length * browser.params.defaultTimeout) : browser.params.defaultTimeout);
+                                   browser.wait(ExpectedConditions.invisibilityOf($('.upload-table')), tableParams.files.length ? (tableParams.records * tableParams.files.length * browser.params.defaultTimeout) : browser.params.defaultTimeout);
                                 }
                                 
                                 // wait for url change
@@ -132,7 +146,10 @@ describe('Record Add', function() {
 					});
 
                     afterAll(function(done) {
-                        tableParams.files.forEach(function(f) {
+                        var files = tableParams.files;
+                        if (process.env.TRAVIS)   files = tableParams.files.filter(function(f) { if (!f.doNotRunInTravis) return f; });
+                        
+                        files.forEach(function(f) {
                             exec('rm ' + f.path);
                         });
                         done();
